Data Analysis Plan
This data analysis plan outlines the methodology for exploring the impact of social support, individual beliefs, and social media use on subjective well-being among emerging adults. Utilising quantitative secondary data analysis (of existing data) methods, the plan outlines the research questions and aims of the study. The analysis will be conducted using RStudio, employing a variety of statistical techniques such as regression analysis (e.g., mediation and interaction effects analysis). The document includes an overview of the pre-existing data selected, from the 'Emerging Adulthood Measured at Multiple Institutions 2: The Next Generation (EAMMi2)' study in 2016/17, the rationale behind the chosen analytic techniques, and a structured approach to exploring the complex interplay between socio-ecological factors and well-being. Designed for flexibility and ongoing learning, this plan serves as a comprehensive guide for examining the nuanced influences on well-being during this critical developmental stage.